Use Map to represent Morphir.SDK.Dict type
A Morphir dict maps arbitrary keys to values. A JavaScript
object can only map strings to values, so it's not quite equivalent
and for this reason a Dict is represented in JSON as an array of
[key,value] pairs.

JavaScript's Map type is equivalent to Morphir Dict and is more
practical than working with an array of pairs directly.